requirements.txt
absl-py==1.4.0
anyio==3.6.2
anykeystore==0.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.8.1
asttokens==2.0.8
astunparse==1.6.3
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==6.0.0
boto3==1.26.59
botocore==1.29.59
bravado==11.0.3
bravado-core==5.17.1
cachetools==5.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.2
contextlib2==21.6.0
contourpy==1.0.5
cryptacular==1.6.2
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
easydict==1.10
executing==1.1.1
faiss-cpu==1.7.4
fastjsonschema==2.16.2
filelock==3.11.0
flatbuffers==23.1.21
fonttools==4.38.0
fqdn==1.5.1
future==0.18.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.3
h5py==3.8.0
hupper==1.11
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipykernel==6.20.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json-tricks==3.17.3
jsonpointer==2.3
jsonref==1.1.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==8.0.1
jupyter-console==6.4.4
jupyter_core==5.1.5
jupyter-events==0.6.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keras==2.11.0
kiwisolver==1.4.4
lazy_loader==0.1
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.1
matplotlib-inline==0.1.6
mistune==2.0.4
monotonic==1.6
msgpack==1.0.4
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
neptune-client==0.16.16
nest-asyncio==1.5.6
networkx==3.0
nni==3.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
nvidia-ml-py==12.535.108
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
PasteDeploy==3.0.1
pbkdf2==1.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
pkgutil_resolve_name==1.3.10
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==2.6.2
prettytable==3.9.0
progressbar2==4.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
pyramid==2.0
pyramid-mailer==0.15.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
python-utils==3.5.2
python3-openid==3.2.0
PythonWebHDFS==0.2.3
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==5.4.1
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
repoze.sendmail==4.4.1
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.23.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987==1.3.8
rsa==4.9
s3transfer==0.6.0
schema==0.7.5
scikit-image==0.20.0
scikit-learn==1.1.2
scipy==1.9.1
Send2Trash==1.8.0
setuptools==63.4.1
simplejson==3.18.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.0
stack-data==0.5.1
swagger-spec-validator==3.0.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
tensorlayer==2.2.5
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.2.28
tinycss2==1.2.1
torch==1.7.1+cu110
torchaudio==0.7.2
torchmetrics==0.10.1
torchvision==0.8.2+cu110
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
transaction==3.0.1
translationstring==1.4
typeguard==4.1.2
types-PyYAML==6.0.12.12
typing_extensions==4.8.0
uri-template==1.2.0
urllib3==1.26.14
velruse==1.1.1
venusian==3.0.0
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.5.0
websockets==11.0.3
Werkzeug==2.2.3
wheel==0.37.1
widgetsnbextension==4.0.5
wrapt==1.15.0
WTForms==3.0.1
wtforms-recaptcha==0.3.2
zipp==3.12.0
zope.deprecation==4.4.0
zope.interface==5.5.2
zope.sqlalchemy==1.6
